Finish first draft of spec/documentation
In this commit, we added lots of text explaining everything.  We also
made this read more like documentation rather than a specification.
However, it's documentation that effectively is a spec and defines the
acceptance criteria.

We also added events related to connecting to and disconnecting from
the CiC global circle.