Update API endpoint and pagination handling
As this API is customer-facing, we can give it a primo URL.  We'll
also take inspiration from Django REST Framework's pagination methods
to improve our approach here.  We'll add a `count` field and return
the first and last time in the resulting data.  We'll also support
parameters for specifying both a closed and open interval.

ref: http://www.django-rest-framework.org/api-guide/pagination/